Ad updates
Added left Ad sidebar. Sidebar is sticky to the top of the viewport on scroll and is only visible on desktop.
Updated right Ad sidebar. Added two more ads and move recommended articles underneath. The last ad is sticky to the top of the viewport, as well as the recommended articles right underneath it.
Added footer Ad bar. Footbar has two ads that wrap when the viewport gets smaller. When it reaches the width of the ads, they disappear.
Left and foot Ad bars are only present in Single template.
Heavily modified Grid layout to accommodate new Ad bars.

Ad bars currently hold placeholders for Ads. That will change with reals ads in the next commit.